Reverted the move of AsyncIter
