+++ tests/unit/v1/test_async_query.py
@@ -17,8 +17,7 @@ import types
 import aiounittest
 
 import mock
-from google.cloud.firestore_v1._helpers import AsyncIter
-from tests.unit.v1.test__helpers import AsyncMock
+from tests.unit.v1.test__helpers import AsyncIter, AsyncMock
 from tests.unit.v1.test_base_query import (
     _make_credentials,
     _make_query_response,
